Task 13: consolidate syncStoreIfBound into shared storeSync.js; replace JSON.parse/stringify with structuredClone
- Create src/services/storeSync.js with createStoreBinding(getState) factory
- Each of the 7 services (hero, dungeon, combat, ui, building, saveLoad,
  production) now calls createStoreBinding() instead of copy-pasting 12 lines
  of boilerplate (let _dispatch + bindStore + syncStoreIfBound)
- structuredClone() replaces JSON.parse(JSON.stringify()) for deep cloning
  (native, faster, handles more types)
