+++ src/services/dungeon.service.js
@@ -2,7 +2,7 @@
  * Dungeon, monster, boss creation, and journey flow.
  * When store is bound via bindStore(store), dispatches REPLACE_STATE after activateDungeon, travel (progress), and monsterFight.
  */
-import { REPLACE_STATE } from '../store/sliceState.js';
+import { createStoreBinding } from './storeSync.js';
 import {
   MONSTERS_PER_BATCH,
   MONSTER_HEALTH_MULTIPLIER,
@@ -14,18 +14,7 @@ import {
 } from '../constants/gameConstants.js';
 
 function DungeonServiceFactory(GameStateService, $timeout, $injector) {
-  var _dispatch = null;
-
-  function bindStore(store) {
-    if (store) _dispatch = store.dispatch;
-  }
-
-  function syncStoreIfBound() {
-    if (_dispatch) {
-      const flat = GameStateService.getState();
-      _dispatch({ type: REPLACE_STATE, payload: JSON.parse(JSON.stringify(flat)) });
-    }
-  }
+  const { bindStore, syncStoreIfBound } = createStoreBinding(() => GameStateService.getState());
 
   function dungeonName() {
     const s = GameStateService.getState();
